refactor: split dashboard form
